requirements.txt
asgiref==3.5.2
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.5
django-filter==22.1
django-phonenumber-field==6.3.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
drf-yasg==1.20.0
freezegun==1.2.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.6.1
MarkupSafe==2.1.1
packaging==21.3
phonenumbers==8.12.50
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-django==4.5.2
pytest-freezegun==0.4.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.9
